Put offer confirmation table into separate file.
