+++ frontend/src/views/assignments/offer-actions.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { connect, useDispatch } from "react-redux";
+import { connect } from "react-redux";
 import { offerTableSelector } from "../offertable/actions";
 import { assignmentsSelector } from "../../api/actions";
 import {
@@ -19,8 +19,7 @@ import {
     FaUserPlus,
 } from "react-icons/fa";
 import { ActionButton } from "../../components/action-buttons";
-import { Button, Modal } from "react-bootstrap";
-import { AdvancedFilterTable } from "../../components/filter-table/advanced-filter-table";
+import { MultiWithdrawOfferConfirmationTable } from "./withdraw-confirmation";
 
 /**
  * Functions to test what actions you can do with a particular assignment
@@ -71,24 +70,8 @@ function OfferActionButtons(props) {
         }
     }
     function confirmOfferWithdraw() {
-        function compareString(str1, str2) {
-            if (str1 > str2) {
-                return 1;
-            } else if (str1 < str2) {
-                return -1;
-            }
-            return 0;
-        }
-
         // if withdrawing multiple offers at once, show confirmation
         if (selectedAssignments?.length > 1) {
-            selectedAssignments.sort((a1, a2) => {
-                return (
-                    compareString(a1.position_code, a2.position_code) ||
-                    compareString(a1.last_name, a2.last_name) ||
-                    compareString(a1.first_name, a2.first_name)
-                );
-            });
             setDdahDeletionConfirmationVisible(true);
         } else {
             // does not need confirmation if only withdrawing one offer
@@ -136,56 +119,6 @@ function OfferActionButtons(props) {
             selectedAssignments.every(OfferTest[key]);
     }
 
-    function MultiWithdrawOfferConfirmationTable() {
-        // const { editable = false } = props;
-        const dispatch = useDispatch();
-        const data = selectedAssignments;
-
-        // We want to minimize the re-render of the table. Since some bindings for columns
-        // are generated on-the-fly, memoize the result so we don't trigger unneeded re-renders.
-        const columns = React.useMemo(() => {
-            return [
-                {
-                    Header: "Last Name",
-                    accessor: "applicant.last_name",
-                },
-                {
-                    Header: "First Name",
-                    accessor: "applicant.first_name",
-                },
-                {
-                    Header: "Position",
-                    accessor: "position.position_code",
-                },
-                {
-                    Header: "Hours",
-                    accessor: "hours",
-                    className: "number-cell",
-                    maxWidth: 70,
-                },
-                {
-                    Header: "Status",
-                    id: "status",
-                    // We want items with no active offer to appear at the end of the list
-                    // when sorted, so we set their accessor to null (the accessor is used by react table
-                    // when sorting items).
-                    accessor: (data) =>
-                        data.active_offer_status === "No Contract"
-                            ? null
-                            : data.active_offer_status,
-                },
-            ];
-        }, [dispatch]);
-
-        return (
-            <AdvancedFilterTable
-                filterable={true}
-                columns={columns}
-                data={data}
-            />
-        );
-    }
-
     return (
         <React.Fragment>
             <ActionButton
@@ -230,38 +163,12 @@ function OfferActionButtons(props) {
             >
                 Set as Rejected
             </ActionButton>
-            <Modal
-                show={ddahDeletionConfirmationVisible}
-                onHide={() => {
-                    setDdahDeletionConfirmationVisible(false);
-                }}
-                size="lg"
-            >
-                <Modal.Header closeButton>
-                    <Modal.Title>Withdrawing Multiple Offers</Modal.Title>
-                </Modal.Header>
-                <Modal.Body>
-                    <div className="mb-3">
-                        You are withdrawing from the following{" "}
-                        {selectedAssignments?.length} offers
-                    </div>
-                    <div className="mb-3">
-                        <MultiWithdrawOfferConfirmationTable />
-                    </div>
-                    Are you sure?
-                </Modal.Body>
-                <Modal.Footer>
-                    <Button
-                        onClick={() => {
-                            setDdahDeletionConfirmationVisible(false);
-                        }}
-                        variant="light"
-                    >
-                        Cancel
-                    </Button>
-                    <Button onClick={withdrawOffers}>Withdraw</Button>
-                </Modal.Footer>
-            </Modal>
+            <MultiWithdrawOfferConfirmationTable
+                data={selectedAssignments}
+                visible={ddahDeletionConfirmationVisible}
+                setVisible={setDdahDeletionConfirmationVisible}
+                withdrawOffers={withdrawOffers}
+            />
         </React.Fragment>
     );
 }
